+++ components/ui/books/club/spreads/poll/poll-page-left.tsx
@@ -16,16 +16,22 @@ export function PollPageLeft({ userSpreadIndex }: Props) {
 	return (
 		<PageLeft userSpreadIndex={userSpreadIndex} id={`club-${pollData?.club_id}-poll-${pollData?.id}-page-left`}>
 			<PollActionsButton />
-			<CardHeader className="px-4 md:px-6 relative h-full pt-4 md:pt-6">
+			<CardHeader className="px-4 md:px-6 relative h-full pt-2 md:pt-6">
 				<CardTitle className="text-md md:text-xl flex flex-row items-center">
 					poll
 					<svg
 						xmlns="http://www.w3.org/2000/svg"
+						fill="none"
 						viewBox="0 0 24 24"
-						fill="currentColor"
-						className="size-4 md:size-6 mx-2"
+						strokeWidth={1.5}
+						stroke="currentColor"
+						className="size-5 md:size-6 mx-2"
 					>
-						<path d="M18.375 2.25c-1.035 0-1.875.84-1.875 1.875v15.75c0 1.035.84 1.875 1.875 1.875h.75c1.035 0 1.875-.84 1.875-1.875V4.125c0-1.036-.84-1.875-1.875-1.875h-.75ZM9.75 8.625c0-1.036.84-1.875 1.875-1.875h.75c1.036 0 1.875.84 1.875 1.875v11.25c0 1.035-.84 1.875-1.875 1.875h-.75a1.875 1.875 0 0 1-1.875-1.875V8.625ZM3 13.125c0-1.036.84-1.875 1.875-1.875h.75c1.036 0 1.875.84 1.875 1.875v6.75c0 1.035-.84 1.875-1.875 1.875h-.75A1.875 1.875 0 0 1 3 19.875v-6.75Z" />
+						<path
+							strokeLinecap="round"
+							strokeLinejoin="round"
+							d="M3 13.125C3 12.504 3.504 12 4.125 12h2.25c.621 0 1.125.504 1.125 1.125v6.75C7.5 20.496 6.996 21 6.375 21h-2.25A1.125 1.125 0 0 1 3 19.875v-6.75ZM9.75 8.625c0-.621.504-1.125 1.125-1.125h2.25c.621 0 1.125.504 1.125 1.125v11.25c0 .621-.504 1.125-1.125 1.125h-2.25a1.125 1.125 0 0 1-1.125-1.125V8.625ZM16.5 4.125c0-.621.504-1.125 1.125-1.125h2.25C20.496 3 21 3.504 21 4.125v15.75c0 .621-.504 1.125-1.125 1.125h-2.25a1.125 1.125 0 0 1-1.125-1.125V4.125Z"
+						/>
 					</svg>
 				</CardTitle>
 				<Separator />