+++ components/ui/books/club/spreads/reading/reading-page-right.tsx
@@ -1,6 +1,6 @@
 "use client"
 
-import { CardFooter, CardHeader, CardTitle } from "@/components/ui"
+import { CardDescription, CardFooter, CardHeader, CardTitle, Separator } from "@/components/ui"
 import { IntervalAvatarGroup, ReadingPosts } from "@/components/ui/books/club/spreads/reading"
 import { CreatePostButton } from "@/components/ui/buttons"
 import { useReading } from "@/contexts"
@@ -23,9 +23,12 @@ export function ReadingPageRight({ userSpreadIndex }: Props) {
 			userSpreadIndex={userSpreadIndex}
 			id={`club-${readingData?.club_id}-reading-${readingData?.id}-page-right`}
 		>
-			<CardHeader className="px-4 md:px-6 h-[calc(100%-6.25rem)] md:h-[calc(100%-118px)] pt-4 md:pt-6">
+			<CardHeader className="px-4 md:px-6 h-[calc(100%-8.5rem)] md:h-[calc(100%-9.625rem)] pt-3 md:pt-6">
+				<CardTitle className="text-md md:text-xl">discussion</CardTitle>
+				<Separator />
+
 				<div className="flex justify-between pr-1">
-					<CardTitle className="text-md md:text-xl">discussion</CardTitle>
+					<CardDescription className="md:text-sm text-xs">share your thoughts!</CardDescription>
 					<CreatePostButton />
 				</div>